+++ src/WebLinter/Constants.cs
@@ -3,5 +3,6 @@
     public static class Constants
     {
         public const string VERSION = "1.0";
+        public const string CACHE_NAME = "WebAnalyzer";
     }
 }