+++ src/WebLinterTest/AssemblyMethods.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using WebLinter;
+
+namespace WebLinterTest
+{
+    [TestClass]
+    public class AssemblyMethods
+    {
+        [AssemblyCleanup]
+        public static void Cleanup()
+        {
+            NodeServer.Down();
+        }
+    }
+}