+++ src/WebLinterVsix/Commnads/CleanErrorsCommand.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.ComponentModel.Design;
 using Microsoft.VisualStudio.Shell;
+using WebLinter;
 
 namespace WebLinterVsix
 {
@@ -49,6 +50,7 @@ namespace WebLinterVsix
         private void CleanErrors(object sender, EventArgs e)
         {
             TableDataSource.Instance.CleanAllErrors();
+            Telemetry.TrackEvent("VS Clean Errors");
         }
     }
 }